Renamed `world` module to `world_data`
